bank/pyproject.toml
[tool.poetry]
name = "bank"
version = "0.1.0"
description = ""
authors = ["Sergey Solop <solop1992@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.12"
flask = "3.0.3"
python-dotenv = "1.0.1"
gunicorn = "23.0.0"
pydantic = "2.9.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
